% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_dashboard.R
\name{get_data_for_dashboard}
\alias{get_data_for_dashboard}
\title{get data for dashboard}
\usage{
get_data_for_dashboard(loc = c("local", "server"))
}
\arguments{
\item{loc}{chr}
}
\value{
list
}
\description{
get data for dashboard
}
