% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\alias{get_all_props}
\title{Get lines and odds on prop bets for popular sports from popular betting sites}
\usage{
get_props(site, sport, prop, save_path = NULL)

get_all_props(
  sports = c("nba"),
  sites = c("dk", "fd", "pb"),
  props = c("ftts", "fpts", "player points alt", "player points ou",
    "player points tiers", "player assists alt", "player assists ou",
    "player assists tiers", "player rebounds alt", "player rebounds ou",
    "player rebounds tiers", "player 3pts alt", "player 3pts ou", "player 3pts tiers"),
  save_path = NULL
)
}
\arguments{
\item{site}{\code{character} book site; support 'fanduel'/'fd', 'draftkings'/'dk', and 'pointsbet'/'pb'}

\item{sport}{\code{character} sport, e.g. 'nba'}

\item{prop}{\code{character} name of prop, e.g. 'first team to score'; see details for available props}

\item{save_path}{\code{character} path to save jsons, if not \code{NULL}}
}
\value{
\code{data.frame} tidy props with tidy columns
}
\description{
Get lines and odds on prop bets for popular sports from popular betting sites
}
\details{
available props include: 'player points alt', 'player points ou', 'player points tiers', 'first team to score', 'first player to score'
}
