% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sportsbook_data.R
\name{get_sportsbook_data}
\alias{get_sportsbook_data}
\title{Get sportsbook data}
\usage{
get_sportsbook_data(
  sportsbook = NA_character_,
  sport = NA_character_,
  save_rds = FALSE,
  save_path = "~/"
)
}
\arguments{
\item{sportsbook}{character name of sports book}

\item{sport}{character sport, which is really the league abbreviation}

\item{save_rds}{logical save off an RDS?}

\item{save_path}{character where to save the file, which takes the format "save_path/sportsbook_sport_datetime.rds"}
}
\value{
list
}
\description{
Get sportsbook data
}
