% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data_for_dashboard.R
\name{output_data_for_dashboard}
\alias{output_data_for_dashboard}
\title{write out the data for the props dashboard}
\usage{
output_data_for_dashboard(
  data_list,
  props_list,
  output_path = system.file("props.rds", package = "betfinder")
)
}
\arguments{
\item{data_list}{list of non-props dfs}

\item{props_list}{list of props dfs}

\item{output_path}{path to save output}
}
\value{
nothing
}
\description{
write out the data for the props dashboard
}
