% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{normalize_names}
\alias{get_key_path}
\alias{hacky_tidyup_player_names}
\title{Utility functions to make things work better}
\usage{
normalize_names(x, key, warn = TRUE)

get_key_path(sport, prop)

hacky_tidyup_player_names(player_names)
}
\arguments{
\item{x}{Vector of strings to normalize in \code{normalize_names}}

\item{key}{List of elements containing fields "name" and "aliases" in \code{normalize_names}}

\item{warn}{Optional, Boolean -- warn if a name in X was not found in the Key names in \code{normalize_names}}
}
\description{
Utility functions to make things work better
}
\note{
\code{normalize_names} was sniped from \code{jimtheflash/gambling_stuff} and written by anthonyshook (to be clear, it was written by a hook owned by some fella named anthony)
}
