% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Get lines on prop bets for popular sports from popular betting sites}
\usage{
get_props(site, sport, prop)
}
\arguments{
\item{site}{character book site; support 'fanduel'/'fd', 'draftkings'/'dk', and 'pointsbet'/'pb'}

\item{sport}{character sport, e.g. 'nba'}

\item{prop}{character name of prop, e.g. 'first team to score'; see details for available props}
}
\value{
data.frame of props with tidy columns
}
\description{
Get lines on prop bets for popular sports from popular betting sites
}
\details{
available props include: 'player points alt', 'player points ou', 'player points tiers', 'first team to score', 'first player to score'
}
