% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Get lines on prop bets for popular sports from popular betting sites}
\usage{
get_props(site, sport, prop, save_path = NULL)
}
\arguments{
\item{site}{\code{character} book site; support 'fanduel'/'fd', 'draftkings'/'dk', and 'pointsbet'/'pb'}

\item{sport}{\code{character} sport, e.g. 'nba'}

\item{prop}{\code{character} name of prop, e.g. 'first team to score'; see details for available props}

\item{save_path}{\code{character} path to save jsons, if not \code{NULL}}
}
\value{
\code{data.frame} tidy props with tidy columns
}
\description{
Get lines on prop bets for popular sports from popular betting sites
}
\details{
available props include: 'player points alt', 'player points ou', 'player points tiers', 'first team to score', 'first player to score'
}
