% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Get odds on prop bets for popular sports from popular betting sites}
\usage{
get_props(site, sport, prop)
}
\arguments{
\item{site}{character site, e.g. 'draftkings' or 'fanduel'}

\item{sport}{character sport, e.g. 'nba'}

\item{prop}{character name of prop, e.g. 'first team to score'}
}
\value{
data.frame of props
}
\description{
Get odds on prop bets for popular sports from popular betting sites
}
