% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{normalize_names}
\alias{get_key_path}
\alias{hacky_tidyup_player_names}
\alias{get_content}
\alias{american_to_prob}
\alias{prob_to_american}
\title{Utility functions to make things work better}
\usage{
normalize_names(x, key, warn = TRUE)

get_key_path(sport, prop)

hacky_tidyup_player_names(player_names)

get_content(uri, query)

american_to_prob(odds)

prob_to_american(probs)
}
\arguments{
\item{x}{Vector of strings to normalize in \code{normalize_names}}

\item{key}{List of elements containing fields "name" and "aliases" in \code{normalize_names}}

\item{warn}{Optional, Boolean -- warn if a name in X was not found in the Key names in \code{normalize_names}}

\item{uri}{main url for call to \code{httr::GET}}

\item{query}{list of paramaters to pass to \code{httr::GET}}
}
\description{
Utility functions to make things work better
}
\note{
\code{normalize_names} was sniped from \code{jimtheflash/gambling_stuff} and written by anthonyshook (to be clear, it was written by a hook owned by some fella named anthony)
}
